+++ nostr/nostr-persona/src/main/java/org/tbk/nostr/persona/Persona.java
@@ -1,4 +1,4 @@
-package org.tbk.nostr.nip19;
+package org.tbk.nostr.persona;
 
 import fr.acinq.bitcoin.Crypto;
 import fr.acinq.bitcoin.MnemonicCode;
@@ -9,7 +9,6 @@ import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
 
-// TODO: move to own package
 public class Persona {
     private Persona() {
         throw new UnsupportedOperationException();